Refactor 言語とタグの型定義をファイル分け
