+++ packages/shared/src/types/categoryTag.ts
@@ -1,0 +1,4 @@
+export type Category = {
+  id: number;
+  category_tag_name: string;
+};
