+++ packages/shared/src/types/communityTag.ts
@@ -1,0 +1,4 @@
+export type CommunityTag = {
+  id: number;
+  community_tag: string;
+};
