+++ packages/shared/src/types/language.ts
@@ -1,0 +1,4 @@
+export type Language = {
+  id: number;
+  language: string;
+};
